+++ DocumentationTests/DocumentationTests/DocumentationTests/04-ui-components/08-participants/02-call-participants.swift
@@ -38,7 +38,6 @@ fileprivate func content() {
             call: viewModel.call,
             participants: viewModel.participants,
             availableFrame: availableFrame,
-            orientation: orientation,
             onChangeTrackVisibility: onChangeTrackVisibility
         )
     }